requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.25.2
pydantic==2.7.4
python-dotenv==1.0.1
loguru==0.7.2
supabase==2.4.0
respx==0.20.2
pytest==8.2.2
pytest-asyncio==0.23.6
